// packages/api-client/src/generated.ts
export interface components {
    schemas: {
        Workflow: {
            /** @constant */
            interfaceVersion: "v1";
            id: string;
            name: string;
            /** Format: date-time */
            createdAt: string;
            start: string;
            steps: ({
                id: string;
                /** @constant */
                type: "task";
                description?: string;
                next?: string;
            } | {
                id: string;
                /** @constant */
                type: "result";
            })[];
        };
        WorkflowPublished: {
            id: string;
            version: number;
            digest: string;
            /** Format: date-time */
            createdAt: string;
            workflow: {
                /** @constant */
                interfaceVersion: "v1";
                id: string;
                name: string;
                /** Format: date-time */
                createdAt: string;
                start: string;
                steps: ({
                    id: string;
                    /** @constant */
                    type: "task";
                    description?: string;
                    next?: string;
                } | {
                    id: string;
                    /** @constant */
                    type: "result";
                })[];
            };
        };
        ErrorResponse: {
            code: string;
            message: string;
            findings: {
                code: string;
                message: string;
                blocking: boolean;
                path: string;
            }[];
        };
        Finding: {
            code: string;
            message: string;
            blocking: boolean;
            path: string;
        };
        Health: {
            /** @constant */
            status: "ok";
        };
        Version: {
            service: string;
            version: string;
            interfaceVersion: string;
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}